Pass the real 16-bit port to port_in/port_out
The callback only ever received the port's low byte, discarding the
high byte (the actual device-select bits) regardless of which OUT/IN
form triggered it. Compute it correctly at every call site instead:
BC for the ED-prefixed IN r,(C)/OUT (C),r forms, and (A<<8)|n for
IN A,(n)/OUT (n),A, matching the real Z80 hardware quirk where A sits
on the upper address lines for that instruction. Without this, OUT
(n),A could never reach any port-decoded device correctly.

+++ z80.h
@@ -10,8 +10,8 @@ typedef struct z80 z80;
 struct z80 {
   uint8_t (*read_byte)(void*, uint16_t);
   void (*write_byte)(void*, uint16_t, uint8_t);
-  uint8_t (*port_in)(z80*, uint8_t);
-  void (*port_out)(z80*, uint8_t, uint8_t);
+  uint8_t (*port_in)(z80*, uint16_t);
+  void (*port_out)(z80*, uint16_t, uint8_t);
   void* userdata;
 
   unsigned long cyc; // cycle count (t-states)
